seperated connection lines into connection.js and added 1 sec timeout to print out main menu
