+++ db/connection.js
@@ -1,0 +1,49 @@
+var mysql = require("mysql");
+var util = require("util");
+var figlet = require("figlet");
+
+///// MYSQL
+var connection = mysql.createConnection({
+  host: "localhost",
+  port: 3306,
+  user: "root",
+  // password: process.env.MYSQL_KEY,
+  password: "password",
+  database: "employee_db",
+});
+
+///// APPEARING TEXT WHEN START APP
+connection.connect(function (err) {
+  if (err) throw err;
+  console.log("connected as id " + connection.threadId + "\n");
+
+  figlet.text(
+    "Welcome To \n Employee \n Manager v1.0",
+    {
+      horizontalLayout: "default",
+      verticalLayout: "full",
+      width: 80,
+      whitespaceBreak: true,
+    },
+    function (err, data) {
+      if (err) {
+        console.log("Something went wrong...");
+        console.dir(err);
+        return;
+      }
+      console.log(
+        "-----------------------------------------------------------------"
+      );
+      console.log(data);
+      console.log(
+        "-----------------------------------------------------------------"
+      );
+      console.log("\n");
+
+    }
+  );
+});
+
+connection.query = util.promisify(connection.query);
+
+module.exports = connection;
