+++ index.js
@@ -1,53 +1,11 @@
 var inquirer = require("inquirer");
-var mysql = require("mysql");
 var figlet = require("figlet");
-var util = require("util");
 var cTable = require("console.table");
+var connection = require("./db/connection");
 
-///// MYSQL
-var connection = mysql.createConnection({
-  host: "localhost",
-  port: 3306,
-  user: "root",
-  // password: process.env.MYSQL_KEY,
-  password: "password",
-  database: "employee_db",
-});
-
-///// APPEARING TEXT WHEN START APP
-connection.connect(function (err) {
-  if (err) throw err;
-  console.log("connected as id " + connection.threadId + "\n");
-
-  figlet.text(
-    "Welcome To \n Employee \n Manager v1.0",
-    {
-      horizontalLayout: "default",
-      verticalLayout: "full",
-      width: 80,
-      whitespaceBreak: true,
-    },
-    function (err, data) {
-      if (err) {
-        console.log("Something went wrong...");
-        console.dir(err);
-        return;
-      }
-      console.log(
-        "-----------------------------------------------------------------"
-      );
-      console.log(data);
-      console.log(
-        "-----------------------------------------------------------------"
-      );
-      console.log("\n");
-
-      mainMenu();
-    }
-  );
-});
-
-connection.query = util.promisify(connection.query);
+function startApp() {
+  setTimeout(mainMenu, 1000);
+}
 
 ///// MAIN MENU (MAIN PROMPT)
 function mainMenu() {
@@ -500,3 +458,5 @@ function endApp() {
   );
   connection.end();
 }
+
+startApp();
